issue 5747 initial refactor
